Fix for Character Controller
CharacterController was removed because it was not used,
CharacterListDBController was renamed to CharacterController,
comments are removed, URI was changed.
