+++ src/main/java/com/springboot/star_wars_character_popularity/app/controller/CharacterController.java
@@ -2,52 +2,75 @@ package com.springboot.star_wars_character_popularity.app.controller;
 
 import com.springboot.star_wars_character_popularity.app.model.Character;
 import com.springboot.star_wars_character_popularity.app.service.CharacterService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.ModelAndView;
+
 import java.util.List;
 
 @RestController
 @RequestMapping("/api/character")
 public class CharacterController {
 
-    @Autowired
     private CharacterService characterService;
 
     public CharacterController(CharacterService characterService){
         this.characterService = characterService;
     }
 
-    @PostMapping("/save")
-    public ModelAndView saveCharacter(@ModelAttribute("character") Character character){
-
-       characterService.saveCharacter(character);
-
-       return new ModelAndView("redirect:/api/characterlistdb/list");
-    }
-
     @GetMapping
-    public List<Character> getAllCharacters(){
-        return characterService.getAllCharacters();
+    public ModelAndView listCharacters(Model theModel){
+
+        List<Character> characters = characterService.getAllCharacters();
+
+        theModel.addAttribute("characters", characters);
+
+        ModelAndView modelAndView = new ModelAndView();
+        modelAndView.setViewName("characters/list-characters");
+
+        return modelAndView;
     }
 
-    @GetMapping("/{id}")
-    public ResponseEntity<Character> getCharacterById(@PathVariable("id") long character_id){
-        return new ResponseEntity<Character>(characterService.getCharacterById(character_id),HttpStatus.OK);
+    @GetMapping("/showFormForAdd")
+    public ModelAndView showFormForAdd(Model theModel){
+
+        Character character = new Character();
+
+        theModel.addAttribute("character", character);
+
+        ModelAndView modelAndView = new ModelAndView();
+        modelAndView.setViewName("characters/character-form");
+
+        return modelAndView;
     }
 
-    @PutMapping("/{id}")
-    public ResponseEntity<Character> updateCharacter(@PathVariable("id") long id, @RequestBody Character character){
-       return new ResponseEntity<Character>(characterService.updateCharacter(character, id), HttpStatus.OK);
+    @GetMapping("/showFormForUpdate")
+    public ModelAndView showFormForUpdate(@RequestParam("characterId") long id, Model theModel){
+
+        Character character = characterService.getCharacterById(id);
+
+        theModel.addAttribute("character", character);
+
+        ModelAndView modelAndView = new ModelAndView();
+        modelAndView.setViewName("characters/character-form");
+
+        return modelAndView;
     }
 
-    @DeleteMapping("/{id}")
-    public ResponseEntity<String> deleteChannel(@PathVariable("id") long id){
+    @GetMapping("/delete")
+    public ModelAndView delete(@RequestParam("characterId") long id){
 
         characterService.deleteCharacter(id);
 
-        return new ResponseEntity<String>("Character deleted successfully !", HttpStatus.OK);
+        return new ModelAndView("redirect:/api/character/");
     }
+
+    @GetMapping("/fetch-characters")
+    public void getCharacterData(){
+        characterService.fetchCharactersOptimized();
+    }
+
 }
